pyproject.toml
[project]
name = "kintone-dify-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
    "tenacity>=9.0.0",
    "uvicorn>=0.32.1",
]
